💄rename : 패키지명 수정
- com.example.travelPlan에서 com.tripfriend로 수정
