+++ backend/src/main/java/com/tripfriend/global/config/SecurityConfig.java
@@ -1,4 +1,4 @@
-package com.example.travelPlan.global.config;
+package com.tripfriend.global.config;
 
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
