+++ backend/src/test/java/com/tripfriend/TravelPlanApplicationTests.java
@@ -1,4 +1,4 @@
-package com.example.travelPlan;
+package com.tripfriend;
 
 import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
